Moved to .NET 5.0, support for VS Code, support for .env file.
