+++ .gitignore
@@ -289,3 +289,4 @@ __pycache__/
 
 /AMSV3Tutorials/packages
 /AMSV3Tutorials/AnalyzeVideos/Output
+/AMSV3Tutorials/.env
