+++ AMSV3Tutorials/AMSV3Tutorials.sln
@@ -13,6 +13,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EncryptWithDRM", "EncryptWi
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Common_Utils", "Common_Utils\Common_Utils.csproj", "{01A8F7CF-C312-494F-90DF-31DF9AD5ACC4}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Live", "Live\Live.csproj", "{1EDEFBD0-86A3-45A5-8A68-CFCD7612B746}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -39,6 +41,10 @@ Global
 		{01A8F7CF-C312-494F-90DF-31DF9AD5ACC4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{01A8F7CF-C312-494F-90DF-31DF9AD5ACC4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{01A8F7CF-C312-494F-90DF-31DF9AD5ACC4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1EDEFBD0-86A3-45A5-8A68-CFCD7612B746}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1EDEFBD0-86A3-45A5-8A68-CFCD7612B746}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1EDEFBD0-86A3-45A5-8A68-CFCD7612B746}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1EDEFBD0-86A3-45A5-8A68-CFCD7612B746}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
