+++ AMSV3Tutorials/Common_Utils/Authentication.cs
@@ -20,6 +20,7 @@ namespace Common_Utils
         /// </summary>
         /// <param name="config">The param is of type ConfigWrapper, which reads values from local configuration file.</param>
         /// <returns>A task.</returns>
+        // <CreateMediaServicesClientAsync>
         public static async Task<IAzureMediaServicesClient> CreateMediaServicesClientAsync(ConfigWrapper config, bool interactive = false)
         {
             ServiceClientCredentials credentials;
@@ -33,6 +34,7 @@ namespace Common_Utils
                 SubscriptionId = config.SubscriptionId,
             };
         }
+        // </CreateMediaServicesClientAsync>
 
         /// <summary>
         /// Create the ServiceClientCredentials object based on the credentials
@@ -65,6 +67,7 @@ namespace Common_Utils
         /// </summary>
         /// <param name="config">The param is of type ConfigWrapper. This class reads values from local configuration file.</param>
         /// <returns></returns>
+        // <GetCredentialsInteractiveAuthAsync>
         private static async Task<ServiceClientCredentials> GetCredentialsInteractiveAuthAsync(ConfigWrapper config)
         {
             var scopes = new[] { config.ArmAadAudience + "/user_impersonation" };
@@ -103,5 +106,6 @@ namespace Common_Utils
 
             return new TokenCredentials(result.AccessToken, TokenType);
         }
+        // </GetCredentialsInteractiveAuthAsync>
     }
 }
