+++ AMSV3Tutorials/EncryptWithDRM/WidevineConfig/OutputProtection.cs
@@ -1,0 +1,23 @@
+﻿using Newtonsoft.Json;
+
+namespace EncryptWithDRM
+{
+    /// <summary>
+    /// OutputProtection Widevine class.
+    /// </summary>
+    public class OutputProtection
+    {
+        /// <summary>
+        /// Gets or sets HDCP protection.
+        /// Supported values : HDCP_NONE, HDCP_V1, HDCP_V2
+        /// </summary>
+        [JsonProperty("hdcp")]
+        public string HDCP { get; set; }
+
+        /// <summary>
+        /// Gets or sets CGMS.
+        /// </summary>
+        [JsonProperty("cgms_flags")]
+        public string CgmsFlags { get; set; }
+    }
+}
