--- AMSV3Tutorials/Live/.gitignore
@@ -1,1 +1,0 @@
-appsettings.json
