+++ AMSV3Tutorials/sample.env
@@ -1,0 +1,26 @@
+# Copy the content of this file to a file named ".env". It should be stored at the root of the repo.
+# Make sure to exclude this file from Git Hub checkins using a .gitignore file
+
+# The values can be obtained from the API Access page for your Media Services account in the portal.
+AADCLIENTID="00000000-0000-0000-0000-000000000000"
+AADSECRET="00000000-0000-0000-0000-000000000000"
+AADTENANTDOMAIN="microsoft.onmicrosoft.com"
+AADTENANTID="00000000-0000-0000-0000-000000000000"
+ACCOUNTNAME="amsaccount"
+RESOURCEGROUP="amsResourceGroup"
+SUBSCRIPTIONID="00000000-0000-0000-0000-000000000000"
+ARMAADAUDIENCE="https://management.core.windows.net/"
+ARMENDPOINT="https://management.azure.com/"
+
+# You must change this if you are using Gov Cloud, China, or other non standard cloud regions
+AADENDPOINT="https://login.microsoftonline.com"
+
+# Settings for the samples
+EVENTHUBCONNECTIONSTRING=""
+EVENTHUBNAME=""
+STORAGECONTAINERNAME=""
+STORAGEACCOUNTNAME=""
+STORAGEACCOUNTKEY=""
+
+# DRM Testing
+SYMMETRICKEY="LPLSvB5RuzFxpVqFxWsUsCL2D83zLOFlGV0R8rTp1f+xrQTWXO/XAQ=="
